# docs/openapi.yaml
openapi: 3.0.0
info:
  version: 0.1.0
  title: SplitNaira API
  description: Premium royalty management API on Stellar network.
servers:
  - url: http://localhost:3001
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT obtained from POST /users/login
    adminApiKey:
      type: apiKey
      in: header
      name: x-admin-api-key
      description: Payments admin API key for /splits/admin mutation endpoints
  schemas:
    ApiError:
      type: object
      properties:
        error:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message
        requestId:
          type: string
          description: Request correlation ID
        details:
          type: object
          additionalProperties:
            nullable: true
          description: Additional error context
      required:
        - error
      example:
        error: VALIDATION_ERROR
        message: collaborators basisPoints must sum to exactly 10000
        requestId: b6f1a2d4-8c3e-4a1b-9f2e-6d5c7a8b9c0d
        details:
          path:
            - collaborators
...
paths:
...
  /users/register:
    post:
      summary: Register a new user
      description: Creates a user profile linked to a Stellar wallet address.
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                walletAddress:
                  type: string
                  minLength: 1
                email:
                  type: string
                  format: email
                alias:
                  type: string
                  minLength: 1
                  maxLength: 64
                role:
                  type: string
                  enum:
                    - company
                    - customer
                    - driver
                  default: customer
              required:
                - walletAddress
      responses:
        "201":
          description: Registered user profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  walletAddress:
                    type: string
                  email:
                    type: string
                  alias:
                    type: string
                  role:
                    type: string
                    enum:
                      - company
                      - customer
                      - driver
                  isActive:
                    type: boolean
                  createdAt:
                    type: string
                  updatedAt:
                    type: string
                required:
                  - id
                  - walletAddress
                  - role
                  - isActive
                  - createdAt
                  - updatedAt
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Conflict with existing resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: CONFLICT
                message: A project with ID "afrobeats_001" already exists.
                requestId: b6f1a2d4-8c3e-4a1b-9f2e-6d5c7a8b9c0d
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: INTERNAL_ERROR
                message: An unexpected error occurred. Please try again later.
                requestId: b6f1a2d4-8c3e-4a1b-9f2e-6d5c7a8b9c0d
  /users/login:
    post:
      summary: Log in by wallet address
      description: Authenticates a registered user and returns a JWT bearer token.
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                walletAddress:
                  type: string
                  minLength: 1
              required:
                - walletAddress
      responses:
        "200":
          description: Authenticated user with JWT
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  walletAddress:
                    type: string
                  email:
                    type: string
                  alias:
                    type: string
                  role:
                    type: string
                    enum:
                      - company
                      - customer
                      - driver
                  isActive:
                    type: boolean
                  createdAt:
                    type: string
                  updatedAt:
                    type: string
                  token:
                    type: string
                required:
                  - id
                  - walletAddress
                  - role
                  - isActive
                  - createdAt
                  - updatedAt
                  - token
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: NOT_FOUND
                message: Project afrobeats_001 does not exist.
                requestId: b6f1a2d4-8c3e-4a1b-9f2e-6d5c7a8b9c0d
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: INTERNAL_ERROR
                message: An unexpected error occurred. Please try again later.
                requestId: b6f1a2d4-8c3e-4a1b-9f2e-6d5c7a8b9c0d
  /users/me:
    get:
      summary: Get authenticated user profile
      description: Returns the profile of the user identified by the JWT bearer token.
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Authenticated user profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  walletAddress:
                    type: string
                  email:
                    type: string
                  alias:
                    type: string
                  role:
                    type: string
                    enum:
                      - company
                      - customer
                      - driver
                  isActive:
                    type: boolean
                  createdAt:
                    type: string
                  updatedAt:
                    type: string
                required:
                  - id
                  - walletAddress
                  - role
                  - isActive
                  - createdAt
                  - updatedAt
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: UNAUTHORIZED
                message: Missing or invalid bearer token / admin API key.
                requestId: b6f1a2d4-8c3e-4a1b-9f2e-6d5c7a8b9c0d
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: NOT_FOUND
                message: Project afrobeats_001 does not exist.
                requestId: b6f1a2d4-8c3e-4a1b-9f2e-6d5c7a8b9c0d
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: INTERNAL_ERROR
                message: An unexpected error occurred. Please try again later.
                requestId: b6f1a2d4-8c3e-4a1b-9f2e-6d5c7a8b9c0d
    patch:
      summary: Update authenticated user profile
      description: Updates email or alias for the authenticated user.
      tags:
        - Users
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                alias:
                  type: string
                  minLength: 1
                  maxLength: 64
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  walletAddress:
                    type: string
                  email:
                    type: string
                  alias:
                    type: string
                  role:
                    type: string
                    enum:
                      - company
                      - customer
                      - driver
                  isActive:
                    type: boolean
                  createdAt:
                    type: string
                  updatedAt:
                    type: string
                required:
                  - id
                  - walletAddress
                  - role
                  - isActive
                  - createdAt
                  - updatedAt
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: UNAUTHORIZED
                message: Missing or invalid bearer token / admin API key.
                requestId: b6f1a2d4-8c3e-4a1b-9f2e-6d5c7a8b9c0d
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: NOT_FOUND
                message: Project afrobeats_001 does not exist.
                requestId: b6f1a2d4-8c3e-4a1b-9f2e-6d5c7a8b9c0d
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: INTERNAL_ERROR
                message: An unexpected error occurred. Please try again later.
                requestId: b6f1a2d4-8c3e-4a1b-9f2e-6d5c7a8b9c0d
  /users/{walletAddress}:
    get:
      summary: Get user by wallet address
      description: Looks up a public user profile by Stellar wallet address.
      tags:
        - Users
      parameters:
        - schema:
            type: string
            minLength: 1
          required: true
          name: walletAddress
          in: path
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  walletAddress:
                    type: string
                  email:
                    type: string
                  alias:
                    type: string
                  role:
                    type: string
                    enum:
                      - company
                      - customer
                      - driver
                  isActive:
                    type: boolean
                  createdAt:
                    type: string
                  updatedAt:
                    type: string
                required:
                  - id
                  - walletAddress
                  - role
                  - isActive
                  - createdAt
                  - updatedAt
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: NOT_FOUND
                message: Project afrobeats_001 does not exist.
                requestId: b6f1a2d4-8c3e-4a1b-9f2e-6d5c7a8b9c0d
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: INTERNAL_ERROR
                message: An unexpected error occurred. Please try again later.
                requestId: b6f1a2d4-8c3e-4a1b-9f2e-6d5c7a8b9c0d
...